requirements.txt
flask==2.3.3
flask-cors==4.0.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
pillow==9.5.0
numpy==1.24.3
torch==2.0.1
torchvision==0.15.2
opencv-python-headless==4.8.1.78
scikit-learn==1.3.0
matplotlib==3.7.2